# src/Api/bin/Debug/net6.0/Swagger/OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pegi",
    "description": "Es una api para la gestion de proyectos de investigación",
    "contact": {
      "name": "Ana Patiño",
      "email": "sofipatino614@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/Proposals": {
      "post": {
        "tags": [
          "Proposal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Proposal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Proposal"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/get-proposals-document/{document}": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/get-proposals-professor/{document}": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/general-statistics-proposal-professor/{document}": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/general-statistics-proposal-student/{document}": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Proposals/get-proposal-code/{code}": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/get-proposals-by-title/{title}": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/update-evaluator-proposal": {
      "put": {
        "tags": [
          "Proposal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/update-tutor-proposal": {
      "put": {
        "tags": [
          "Proposal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Proposals/{code}": {
      "delete": {
        "tags": [
          "Proposal"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProposalRequest": {
        "type": "object",
        "properties": {
          "personDocument1": {
            "type": "string",
            "nullable": true
          },
          "personDocument2": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "investigationGroup": {
            "type": "string",
            "nullable": true
          },
          "approach": {
            "type": "string",
            "nullable": true
          },
          "justification": {
            "type": "string",
            "nullable": true
          },
          "generalObjective": {
            "type": "string",
            "nullable": true
          },
          "specificObjective": {
            "type": "string",
            "nullable": true
          },
          "bibliographical": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "thematicAreaCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProposalUpdate": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "personDocument1": {
            "type": "string",
            "nullable": true
          },
          "personDocument2": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "investigationGroup": {
            "type": "string",
            "nullable": true
          },
          "approach": {
            "type": "string",
            "nullable": true
          },
          "justification": {
            "type": "string",
            "nullable": true
          },
          "generalObjective": {
            "type": "string",
            "nullable": true
          },
          "specificObjective": {
            "type": "string",
            "nullable": true
          },
          "bibliographical": {
            "type": "string",
            "nullable": true
          },
          "thematicAreaCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProposalUpdateRequest": {
        "type": "object",
        "properties": {
          "professorDocument": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}